Was passing port obtained by querying grid service as int instead of unsigned short int - caused value to be lost (possibly -O2 specific) and replaced with zero.

+++ include/ReG_Steer_Appside_Soap.h
@@ -52,8 +52,8 @@ struct msg_struct *Get_control_msg_soap();
 
 int Finalize_steering_connection_soap();
 
-int Get_data_source_address_soap(int   index, 
-				 char *hostname,
-				 int  *port);
+int Get_data_source_address_soap(int                 index, 
+				 char               *hostname,
+				 unsigned short int *port);
 
 #endif
